Vite Larn 12.6
Based on https://github.com/primeau/Larn

The build didn't work for me, so I forked it and changed it to use
[Vite](https://vitejs.dev/).  There are some things to access Amazon
AWS for some things (scoreboard?  saving?) that I didn't care about
since I just wanted to play the game so they probably aren't working.
I also added cheats that can be accessed in the browser console
on the global cheats object.