Update object.js to fix 1 charater for fortune cookie hint (change E to e)
Hello!  Thanks for Larn, I've been having some fun playing it and noticed a minor error.  I'm new to GitHub so I hope it is OK for me to propose a PR like this, and that I get things correct :)

This PR fixes a one-character typo (should be lower case 'e' to eat) in src/object.js line 757:

- if (do_ident) updateLog(`${youFound} ${item}${period}`, formatHint('t', 'to take', 'E', 'to eat'));
+ if (do_ident) updateLog(`${youFound} ${item}${period}`, formatHint('t', 'to take', 'e', 'to eat'));

Thanks for making the source code public!  Cheers and happy gaming, John